feat: Set volume size using snapshot options
Support setting a custom ZFS volume size for snapshots.

The volume size can be configured by appending the label
'containerd.io/snapshot/zvol/size' with the volume size in bytes as the
value.

The value can not be smaller than the configured default volume size and
has to be larger or equal to the volume size of its parent.

The helper function 'WithVolumeSize' can be used to set the label.

Signed-off-by: Han Verstraete <welteki@pm.me>